split component-filters to individual comps
