+++ scopes/component/component-filters/deprecate-filter/deprecate-filter.module.scss
