+++ scopes/component/component-filters/deprecate-filter/index.ts
@@ -1,0 +1,1 @@
+export { DeprecateFilter } from './deprecate-filter';
