+++ scopes/component/component-filters/env-filter/index.ts
@@ -1,0 +1,1 @@
+export { EnvsFilter } from './envs-filter';
