--- scopes/component/component-filters/index.ts
@@ -1,9 +1,0 @@
-export {
-  ComponentFiltersProvider,
-  ComponentFilterContext,
-  ComponentFilterCriteria,
-  ComponentFilters,
-  runAllFilters,
-} from './component-filters.context';
-export { DeprecateFilter } from './deprecate-filter';
-export { EnvsFilter } from './envs-filter';
