+++ scopes/component/component-filters/use-component-filter/index.ts
@@ -1,0 +1,8 @@
+export {
+  ComponentFiltersProvider,
+  ComponentFilterContext,
+  ComponentFilterCriteria,
+  ComponentFilters,
+  runAllFilters,
+  useComponentFilter,
+} from './use-component-filter';
